requirements.txt
redis>=2.0
PyYAML>=3.11
mockredispy>=2.9.0.0
